+++ src/lazy-checked-clone.js
@@ -23,11 +23,8 @@
 let spec = require('shift-spec').default;
 let Shift = require('shift-ast/checked');
 
+let { isNodeOrUnionOfNodes, isStatefulType } = require('./utilities.js');
 
-// TODO consider splitting these and the utilities from https://github.com/shapesecurity/shift-reducer-js/blob/es2016/scripts/lib/utilities.js into their own project
-function isNodeOrUnionOfNodes(type) {
-  return type.typeName === 'Union' && type.arguments.every(isNodeOrUnionOfNodes) || spec.hasOwnProperty(type.typeName);
-}
 
 function equals(type, a, b) {
   switch (type.typeName) {
@@ -66,26 +63,6 @@ function equals(type, a, b) {
         return a === b;
       }
       throw new Error('unimplemented: ' + type);
-  }
-}
-
-function isStatefulType(type) {
-  switch (type.typeName) {
-    case 'Enum':
-    case 'String':
-    case 'Number':
-    case 'Boolean':
-      return false;
-    case 'Maybe':
-    case 'List':
-      return isStatefulType(type.argument);
-    case 'Union':
-      return type.arguments.some(isStatefulType);
-    default:
-      if (isNodeOrUnionOfNodes(type)) {
-        return true;
-      }
-      throw new Error('unimplemented: type ' + type);
   }
 }
